Move all end-to-end tests to the tests folder
To better align with the golang project recommended layout moving all non-unit
tests to the /tests folder. These require full e2e running of containers in
various ways. All unit test's should still be written inline with the package
under test as normal.
