--- functional/manifest.go
@@ -1,3 +1,0 @@
-package functional
-
-import _ "github.com/Microsoft/hcsshim/functional/manifest"
