--- pkg/go-runhcs/runhcs_integration_test.go
@@ -1,7 +1,0 @@
-// +build integration
-
-package runhcs
-
-import (
-	_ "github.com/Microsoft/hcsshim/functional/manifest"
-)
