+++ test/functional/assets/defaultlinuxspec.json
