+++ test/functional/assets/defaultwindowsspec.json
