+++ test/functional/assets/samples/from-docker-linux/privileged.json
