+++ test/functional/assets/samples/from-docker-linux/sh.json
