+++ test/functional/manifest/manifest.go
