+++ test/functional/manifest/rsrc_amd64.syso
