+++ test/functional/manifest_test.go
@@ -1,0 +1,3 @@
+package functional
+
+import _ "github.com/Microsoft/hcsshim/test/functional/manifest"
