+++ test/functional/utilities/scratch.go
