+++ test/functional/wcow_xenon_v2_test.go
@@ -6,7 +6,7 @@ package functional
 //	"os"
 //	"testing"
 
-//	"github.com/Microsoft/hcsshim/functional/utilities"
+//	"github.com/Microsoft/hcsshim/test/functional/utilities"
 //	"github.com/Microsoft/hcsshim/internal/guid"
 //	"github.com/Microsoft/hcsshim/internal/hcsoci"
 //	"github.com/Microsoft/hcsshim/osversion"
