+++ test/runhcs/runhcs_test.go
@@ -1,0 +1,7 @@
+// +build integration
+
+package runhcs
+
+import (
+	_ "github.com/Microsoft/hcsshim/test/functional/manifest"
+)
